requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.17
openai==1.77.0
faiss-cpu==1.13.2
sentence-transformers==4.1.0
pypdf==5.4.0
python-docx==1.1.2
docx2txt==0.9
python-dotenv==1.1.0
tiktoken==0.9.0
streamlit==1.44.1
python-multipart==0.0.20
